Move view classes to models and controllers
