+++ common/app/common/controllers/PageData.java
@@ -1,4 +1,4 @@
-package common.pages;
+package common.controllers;
 
 public interface PageData {
