+++ common/app/common/models/NavMenuData.java
@@ -1,7 +1,7 @@
 package common.models;
 
 import common.contexts.UserContext;
-import common.pages.ReverseRouter;
+import common.controllers.ReverseRouter;
 import io.sphere.sdk.categories.Category;
 import io.sphere.sdk.categories.CategoryTree;
 import io.sphere.sdk.models.Base;
