+++ product-catalog/app/productcatalog/models/SelectorData.java
@@ -1,4 +1,4 @@
-package productcatalog.pages;
+package productcatalog.models;
 
 import common.models.SelectableData;
