+++ product-catalog/app/productcatalog/models/SortSelector.java
@@ -1,4 +1,4 @@
-package productcatalog.pages;
+package productcatalog.models;
 
 import io.sphere.sdk.models.Base;
 import io.sphere.sdk.products.ProductProjection;
