+++ purchase/app/purchase/CartDetailPageController.java
@@ -3,7 +3,7 @@ package purchase;
 import common.contexts.UserContext;
 import common.controllers.ControllerDependency;
 import common.models.ProductDataConfig;
-import common.pages.SunrisePageData;
+import common.controllers.SunrisePageData;
 import io.sphere.sdk.carts.Cart;
 import io.sphere.sdk.carts.commands.CartUpdateCommand;
 import io.sphere.sdk.carts.commands.updateactions.AddLineItem;
