+++ purchase/app/purchase/CheckoutPaymentPageContent.java
@@ -2,7 +2,7 @@ package purchase;
 
 import common.contexts.UserContext;
 import common.models.ProductDataConfig;
-import common.pages.ReverseRouter;
+import common.controllers.ReverseRouter;
 import io.sphere.sdk.carts.Cart;
 
 public class CheckoutPaymentPageContent extends CheckoutPageContent {
