+++ purchase/app/purchase/CheckoutShippingFormBean.java
@@ -1,7 +1,7 @@
 package purchase;
 
 import common.contexts.UserContext;
-import common.pages.ReverseRouter;
+import common.controllers.ReverseRouter;
 import io.sphere.sdk.carts.Cart;
 import io.sphere.sdk.models.Address;
 import play.Configuration;
